requirements.txt
adal==1.2.7
asgiref==3.4.1
astroid==2.9.3
autopep8==1.5.7
azure==4.0.0
azure-applicationinsights==0.1.0
azure-batch==4.1.3
azure-common==1.1.27
azure-core==1.23.1
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-datalake-store==0.0.52
azure-eventgrid==1.3.0
azure-graphrbac==0.40.0
azure-identity==1.8.0
azure-keyvault==1.1.0
azure-loganalytics==0.1.1
azure-mgmt==4.0.0
azure-mgmt-advisor==1.0.1
azure-mgmt-applicationinsights==0.1.1
azure-mgmt-authorization==0.50.0
azure-mgmt-batch==5.0.1
azure-mgmt-batchai==2.0.0
azure-mgmt-billing==0.2.0
azure-mgmt-cdn==3.1.0
azure-mgmt-cognitiveservices==3.0.0
azure-mgmt-commerce==1.0.1
azure-mgmt-compute==4.6.2
azure-mgmt-consumption==2.0.0
azure-mgmt-containerinstance==1.5.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-containerservice==4.4.0
azure-mgmt-cosmosdb==0.4.1
azure-mgmt-datafactory==0.6.0
azure-mgmt-datalake-analytics==0.6.0
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==1.0.0
azure-mgmt-devspaces==0.1.0
azure-mgmt-devtestlabs==2.2.0
azure-mgmt-dns==2.1.0
azure-mgmt-eventgrid==1.0.0
azure-mgmt-eventhub==2.6.0
azure-mgmt-hanaonazure==0.1.1
azure-mgmt-iotcentral==0.1.0
azure-mgmt-iothub==0.5.0
azure-mgmt-iothubprovisioningservices==0.2.0
azure-mgmt-keyvault==1.1.0
azure-mgmt-loganalytics==0.2.0
azure-mgmt-logic==3.0.0
azure-mgmt-machinelearningcompute==0.4.1
azure-mgmt-managementgroups==0.1.0
azure-mgmt-managementpartner==0.1.1
azure-mgmt-maps==0.1.0
azure-mgmt-marketplaceordering==0.1.0
azure-mgmt-media==1.0.1
azure-mgmt-monitor==0.5.2
azure-mgmt-msi==0.2.0
azure-mgmt-network==2.7.0
azure-mgmt-notificationhubs==2.1.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==0.1.0
azure-mgmt-powerbiembedded==2.0.0
azure-mgmt-rdbms==1.9.0
azure-mgmt-recoveryservices==0.3.0
azure-mgmt-recoveryservicesbackup==0.3.0
azure-mgmt-redis==5.0.0
azure-mgmt-relay==0.1.0
azure-mgmt-reservations==0.2.1
azure-mgmt-resource==2.2.0
azure-mgmt-scheduler==2.0.0
azure-mgmt-search==2.1.0
azure-mgmt-servicebus==0.5.3
azure-mgmt-servicefabric==0.2.0
azure-mgmt-signalr==0.1.1
azure-mgmt-sql==0.9.1
azure-mgmt-storage==2.0.0
azure-mgmt-subscription==0.2.0
azure-mgmt-trafficmanager==0.50.0
azure-mgmt-web==0.35.0
azure-nspkg==3.0.2
azure-servicebus==0.21.1
azure-servicefabric==6.3.0.0
azure-servicemanagement-legacy==0.20.7
azure-storage-blob==1.5.0
azure-storage-common==1.4.2
azure-storage-file==1.4.0
azure-storage-queue==1.4.0
bandit==1.7.1
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
Django==4.0.1
django-cors-headers==3.11.0
django-extensions==3.1.5
django-filter==21.1
django-regex==0.4.1
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
dodgy==0.2.1
drf-yasg==1.20.0
flake8==3.9.2
flake8-polyfill==1.0.2
gitdb==4.0.9
GitPython==3.1.26
google-api-core==2.7.1
google-auth==2.6.2
googleapis-common-protos==1.56.0
idna==3.3
importlib-metadata==4.10.1
inflection==0.5.1
isodate==0.6.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
lazy-object-proxy==1.7.1
Markdown==3.3.6
MarkupSafe==2.0.1
mccabe==0.6.1
msal==1.17.0
msal-extensions==0.3.1
msrest==0.6.21
msrestazure==0.6.4
mssql-django==1.1.2
oauthlib==3.1.1
opencensus==0.8.0
opencensus-context==0.1.2
opencensus-ext-azure==1.1.3
packaging==21.3
pbr==5.8.0
pep8-naming==0.10.0
platformdirs==2.4.1
portalocker==2.4.0
prospector==1.6.0
protobuf==3.20.0
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.21
pydocstyle==6.1.1
pyflakes==2.3.1
PyJWT
pylint==2.12.2
pylint-celery==0.3
pylint-django==2.5.0
pylint-flask==0.6
pylint-plugin-utils==0.7
pyparsing==3.0.7
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.0
requirements-detector==0.7
rsa==4.8
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
setoptconf-tmp==0.3.1
simplejson==3.17.6
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
sqlparse==0.4.2
stevedore==3.5.0
toml==0.10.2
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
wrapt==1.13.3
zipp==3.7.0